Add files via upload
employee/manager/customer_Settings.html: added a submit button and changed the placeholders for entering a new password

customer_view.html: You can select a hairstyle and the value will be stored in a (currently visible) HTML tag. I need to continue working on the Select Stylist box's JavaScript so that the selected stylist can also be stored in an HTML tag. Also need to finish the CANCEL button on appointments.

validation.js: added functions to store selected hairstyle and stylist (need to finish the stylist function -- line 931). Removed extra alerts in the file.